pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}  # Map the root to "src"

[tool.setuptools.packages.find]
where = ["src"]

[project]
name = "xvault"
version = "0.0.35"
license = "MIT"
authors = [
  { name="Marc Delos", email="marcdp@dprojects.com" },
]
dependencies = [
    "dprojectstools>=0.0.29",
    "argon2-cffi>=25.1.0",
    "python-dotenv>=1.2.2",
    "keyring>=25.6.0",
    "cryptography>=44.0.0",
    "pyperclip>=1.9.0",
    "lxml>=5.3.0",
    "json5>=0.13.0"
]
description = "Secure storage vault for secrets in git repositories (SOPS alike)"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]

[project.scripts]
xvault = "xvault.__main__:main"

[project.urls]
Homepage = "https://github.com/marcdp/xvault"
Issues = "https://github.com/marcdp/xvault/issues"
